Make the performance dashboard no longer public
