+++ app/controllers/support_interface/performance_dashboard_controller.rb
@@ -1,5 +1,5 @@
-module Integrations
-  class PerformanceDashboardController < ApplicationController
+module SupportInterface
+  class PerformanceDashboardController < SupportInterfaceController
     def dashboard
       raise ArgumentError unless params[:year].in?([nil, '2020', '2021'])
