+++ app/views/support_interface/performance/index.html.erb
@@ -1,7 +1,7 @@
 <%= render 'performance_navigation', title: 'Dashboards' %>
 
 <ul class="govuk-list govuk-list--bullet">
-  <li><%= govuk_link_to 'Service performance', integrations_performance_path %> (public)</li>
+  <li><%= govuk_link_to 'Service performance', support_interface_performance_dashboard_path %></li>
   <li><%= govuk_link_to 'Course stats', support_interface_course_stats_path %></li>
   <li><%= govuk_link_to 'Course options without vacancies', support_interface_course_options_path %></li>
   <li><%= govuk_link_to 'Applications with unavailable choices', support_interface_unavailable_choices_path %></li>
